Add Wifi Connection Files to Folder
The arduino script connects to the "DukeOpen" network. For the board to connect, its MAC address must be registered through "DukeReg". The MAC address can be obtained through the "ScanNetworks" example file under the "WIFI101" library. The SSID and Password for network connections are entered through the "arduino_secrets.h" file.